requirements.txt
peewee==3.16.0
python-dotenv==1.00.0
uuid==1.30
click==8.1.3
mariadb==1.1.6